refactor: 스타일 컴포넌트 파일 분리
- 스타일 코드가 길어질수록 컴포넌트에 대한 로직이 잘 보이지 않아 스타일 코드 파일 분리 작업을 진행했습니다.
- 일부 파일은 전체적인 로직 변경필요해서 아직 리팩토링 진행 X
