+++ fe/src/components/common/header/Header.jsx
@@ -1,17 +1,6 @@
-import styled from 'styled-components';
 import Logo from '@assets/images/Logo.svg';
 import UserImageLarge from '@assets/images/UserImageLarge.svg';
-
-const HeaderContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 50px;
-`;
-
-const UserImage = styled.img`
-  width: 32px;
-  height: 32px;
-`;
+import { HeaderContainer, UserImage } from './HeaderStyled';
 
 const Header = () => {
   return (
