+++ fe/src/components/common/header/HeaderStyled.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const HeaderContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 50px;
+`;
+
+export const UserImage = styled.img`
+  width: 32px;
+  height: 32px;
+`;
