+++ fe/src/components/common/tab/Tab.jsx
@@ -1,16 +1,5 @@
-import styled from 'styled-components';
 import Button from '../button/Button';
-
-const TabContainer = styled.div`
-  flex-direction: row;
-  justify-content: space-evenly;
-  align-items: center;
-  padding: 0px 16px;
-  height: 40px;
-  display: flex;
-  border-radius: 11px;
-  border: 1px solid #d9dbe9;
-`;
+import TabContainer from './TabStyled';
 
 const Tab = () => {
   return (
