+++ fe/src/components/common/tab/TabStyled.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const TabContainer = styled.div`
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+  padding: 0px 16px;
+  height: 40px;
+  display: flex;
+  border-radius: 11px;
+  border: 1px solid #d9dbe9;
+`;
+
+export default TabContainer;
